requirements.txt
flask
tensorflow
nltk
pandas
scikit-learn
matplotlib
haversine
ortools
gunicorn
google-cloud-firestore
